refactor deploy dcnt & lock release with extra comments
