+++ .env.example
@@ -1,3 +1,4 @@
-ETHERSCAN_API_KEY=ABC123ABC123ABC123ABC123ABC123ABC1
-ROPSTEN_URL=https://eth-ropsten.alchemyapi.io/v2/<YOUR ALCHEMY KEY>
-PRIVATE_KEY=0xabc123abc123abc123abc123abc123abc123abc123abc123abc123abc123abc1
+ETHERSCAN_API_KEY=
+MAINNET_PROVIDER=
+GOERLI_PROVIDER=
+DEPLOYER_PRIVATE_KEY=
