+++ gas-report.txt
@@ -11,7 +11,7 @@
 ················|············|··············|·············|·············|···············|··············
 |  DCNTToken    ·  mint      ·       73486  ·      79182  ·      76870  ·            5  ·          -  │
 ················|············|··············|·············|·············|···············|··············
-|  DCNTToken    ·  transfer  ·       52017  ·     123333  ·      69955  ·            7  ·          -  │
+|  DCNTToken    ·  transfer  ·       52029  ·     123333  ·      63919  ·           13  ·          -  │
 ················|············|··············|·············|·············|···············|··············
 |  LockRelease  ·  delegate  ·      102516  ·     102528  ·     102522  ·            2  ·          -  │
 ················|············|··············|·············|·············|···············|··············
@@ -19,7 +19,7 @@
 ················|············|··············|·············|·············|···············|··············
 |  Deployments               ·                                          ·  % of limit   ·             │
 ·····························|··············|·············|·············|···············|··············
-|  DCNTToken                 ·     4297363  ·    4297447  ·    4297390  ·       14.3 %  ·          -  │
+|  DCNTToken                 ·     4297363  ·    4297447  ·    4297392  ·       14.3 %  ·          -  │
 ·····························|··············|·············|·············|···············|··············
-|  LockRelease               ·           -  ·          -  ·    3976641  ·       13.3 %  ·          -  │
+|  LockRelease               ·     3723633  ·    3976641  ·    3933268  ·       13.1 %  ·          -  │
 ·----------------------------|--------------|-------------|-------------|---------------|-------------·
