+++ scripts/deploy.ts
@@ -1,28 +1,161 @@
-import ethProvider from "eth-provider";
+import { AzoriusTxBuilder } from "../DaoBuilder/AzoriusTxBuilder";
 import { ethers } from "hardhat";
 
-async function main() {
-  const frame = ethProvider("frame");
+import { getMasterCopies, getSafeData } from "../DaoBuilder/daoUtils";
+import { deployDCNTAndLockRelease } from "../DaoBuilder/dcntTokenDeploy";
+import { encodeMultiSend } from "../DaoBuilder/utils";
+import { decentDAOConfig } from "../DaoBuilder/dcntDAOConfig";
+import { utils } from "ethers";
+import { logDcntDaoTxt } from "./graphics/graphics";
 
-  const freeMintWhole = 1_000_000_000; // 1 billion
-  const freeMintTotal = ethers.utils.parseEther(freeMintWhole.toString());
+async function createDAO() {
 
-  const DCNTToken = await ethers.getContractFactory("DCNTToken");
-  const dcntTx = await DCNTToken.getDeployTransaction(
-    freeMintTotal,
-    frame.address
+  // Deploy DCNT Token
+  const [deployer] = await ethers.getSigners();
+  const { dcntTokenContract, lockReleaseContract, totalLockedAmount } =
+    await deployDCNTAndLockRelease(deployer, decentDAOConfig);
+
+  console.log("Decent Token and Lock Release contracts deployed");
+  console.table({
+    dcntTokenContract: dcntTokenContract.address,
+    lockReleaseContract: lockReleaseContract.address,
+    totalLockedAmount: totalLockedAmount.toString(),
+  });
+
+  const {
+    zodiacModuleProxyFactoryContract,
+    fractalAzoriusMasterCopyContract,
+    fractalRegistryContract,
+    keyValuePairContract,
+    linearVotingMasterCopyContract,
+    multisendContract,
+  } = await getMasterCopies();
+
+  const { predictedSafeContract, createSafeTx } = await getSafeData(
+    multisendContract
   );
 
-  const deployer: string = (
-    await frame.request({ method: "eth_requestAccounts" })
-  )[0];
+  console.log("Master copies fetched");
+  console.table({
+    zodiacModuleProxyFactoryContract: zodiacModuleProxyFactoryContract.address,
+    fractalAzoriusMasterCopyContract: fractalAzoriusMasterCopyContract.address,
+    fractalRegistryContract: fractalRegistryContract.address,
+    keyValuePairContract: keyValuePairContract.address,
+    linearVotingMasterCopyContract: linearVotingMasterCopyContract.address,
+    multisendContract: multisendContract.address,
+  });
 
-  dcntTx.from = deployer;
+  const azoriusTxBuilder = new AzoriusTxBuilder(
+    decentDAOConfig,
+    predictedSafeContract,
+    dcntTokenContract,
+    lockReleaseContract,
+    multisendContract,
+    zodiacModuleProxyFactoryContract,
+    fractalAzoriusMasterCopyContract,
+    fractalRegistryContract,
+    keyValuePairContract,
+    linearVotingMasterCopyContract
+  );
 
-  await frame.request({ method: "eth_sendTransaction", params: [dcntTx] });
+  console.log("Azorius tx builder created");
+
+  const txs = [createSafeTx];
+  const internalTxs = [
+    azoriusTxBuilder.buildUpdateDAONameTx(),
+    azoriusTxBuilder.buildUpdateDAOSnapshotURLTx(),
+    azoriusTxBuilder.buildLinearVotingContractSetupTx(),
+    azoriusTxBuilder.buildEnableAzoriusModuleTx(),
+    azoriusTxBuilder.buildAddAzoriusContractAsOwnerTx(),
+    azoriusTxBuilder.buildRemoveMultiSendOwnerTx(),
+  ];
+  console.log("Internal txs created");
+
+  txs.push(azoriusTxBuilder.buildDeployStrategyTx());
+  txs.push(azoriusTxBuilder.buildDeployAzoriusTx());
+  txs.push(
+    azoriusTxBuilder.buildExecInternalSafeTx(
+      azoriusTxBuilder.signatures(),
+      internalTxs
+    )
+  );
+  console.log("Tx created");
+
+  const encodedTx = encodeMultiSend(txs);
+
+  console.time("Tx sent");
+  const execution = await multisendContract.multiSend(encodedTx, {
+    gasLimit: 5000000,
+  });
+  execution.wait();
+  console.timeEnd(`tx executed ${execution.hash}`);
+
+  logDcntDaoTxt();
+  console.table({ daoAddress: predictedSafeContract.address });
+
+  console.log("Transferring remaining balance of DCNT to Decent DAO");
+  const amountToTransfer = utils
+    .parseEther(decentDAOConfig.initialSupply)
+    .sub(totalLockedAmount);
+  const tokenTransfer = await dcntTokenContract.transfer(
+    predictedSafeContract.address,
+    amountToTransfer
+  );
+  tokenTransfer.wait();
+  console.log("Tokens transferred");
+
+  console.table({
+    amountToTransfer: amountToTransfer,
+    hash: tokenTransfer.hash,
+  });
+
+  console.log("Transferring token ownership to Decent DAO");
+  const transferTokenOwnership = await dcntTokenContract.transferOwnership(
+    predictedSafeContract.address
+  );
+  transferTokenOwnership.wait();
+  console.log("Token ownership transferred");
+  console.table({
+    dao: predictedSafeContract.address,
+    hash: transferTokenOwnership.hash,
+  });
+  console.log(`
+      ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⣠⣤⣴⣶⣶⣿⣿⣿⣿⣿⣿⣿⣿⣶⣶⣦⣤⣄⣀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
+      ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣠⣴⣾⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣷⣦⣄⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
+      ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣤⣾⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣷⣤⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
+      ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣠⣾⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡟⢿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣷⣄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
+      ⠀⠀⠀⠀⠀⠀⠀⠀⣠⣾⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⠏⠀⠀⠻⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣷⣄⠀⠀⠀⠀⠀⠀⠀⠀
+      ⠀⠀⠀⠀⠀⠀⣠⣾⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⠋⠀⠀⠀⠀⠹⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣷⣄⠀⠀⠀⠀⠀⠀
+      ⠀⠀⠀⠀⠀⣴⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡿⠁⠀⠀⠀⠀⠀⠀⠘⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣦⠀⠀⠀⠀⠀
+      ⠀⠀⠀⢀⣾⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡟⠁⠀⠀⠀⠀⠀⠀⠀⠀⠈⢿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣷⡀⠀⠀⠀
+      ⠀⠀⢀⣾⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⠟⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢻⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣷⡀⠀⠀
+      ⠀⠀⣾⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⠋⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠹⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣷⠀⠀
+      ⠀⣸⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡿⠃⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠘⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣇⠀
+      ⢀⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡿⠁⠀⠀⠀⠀⠀⠀⠀⠀⣀⣀⠀⠀⠀⠀⠀⠀⠀⠀⠈⢿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡀
+      ⢸⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⠟⠀⠀⠀⠀⠀⢀⣠⣴⣶⠿⠟⠻⢿⣶⣤⣄⡀⠀⠀⠀⠀⠈⢻⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡇
+      ⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⠏⠀⠀⣀⣤⣶⡾⠿⠛⠉⠀⠀⠀⠀⠀⠀⠉⠛⠿⣷⣦⣄⣀⠀⠀⠻⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣷
+      ⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣣⣤⣶⠿⠟⠋⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⠙⠻⢿⣶⣤⣹⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿
+      ⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡛⠉⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⠉⣻⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿
+      ⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡿⣷⣦⣄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣠⣴⣿⢿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡿
+      ⢸⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣄⠉⠻⢿⣶⣤⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⣤⣾⡿⠛⠉⣠⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡇
+      ⠈⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣦⡀⠀⠈⠙⠿⣷⣦⣄⠀⠀⠀⠀⠀⠀⣠⣴⣾⠟⠋⠁⠀⢀⣼⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⠁
+      ⠀⢹⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣷⡄⠀⠀⠀⠀⠙⠻⢿⣶⣄⣤⣶⡿⠟⠉⠀⠀⠀⠀⣠⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡏⠀
+      ⠀⠀⢿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣦⠀⠀⠀⠀⠀⠀⠈⠛⠋⠁⠀⠀⠀⠀⠀⢀⣴⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡿⠀⠀
+      ⠀⠀⠈⢿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣷⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣠⣾⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡿⠁⠀⠀
+      ⠀⠀⠀⠈⢿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣆⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣴⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡿⠁⠀⠀⠀
+      ⠀⠀⠀⠀⠀⠻⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣷⡀⠀⠀⠀⠀⠀⠀⢠⣾⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⠟⠀⠀⠀⠀⠀
+      ⠀⠀⠀⠀⠀⠀⠙⢿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣄⠀⠀⠀⠀⣰⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡿⠋⠀⠀⠀⠀⠀⠀
+      ⠀⠀⠀⠀⠀⠀⠀⠀⠙⢿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣧⡀⢀⣾⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡿⠋⠀⠀⠀⠀⠀⠀⠀⠀
+      ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⢿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡿⠋⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
+      ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⠛⢿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡿⠛⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
+      ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⠙⠻⢿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⡿⠟⠋⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
+      ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠉⠙⠛⠻⠿⠿⢿⣿⣿⣿⣿⣿⣿⡿⠿⠿⠟⠛⠋⠉⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
+      `);
 }
 
-main().catch((error) => {
-  console.error(error);
-  process.exitCode = 1;
-});
+createDAO()
+  .then(() => process.exit(0))
+  .catch((error) => {
+    console.error(error);
+    process.exit(1);
+  });
