+++ scripts/graphics/graphics.ts
@@ -1,0 +1,10 @@
+export const logDcntDaoTxt = () => {
+  console.log(`
+  888~-_                                          d8         888~-_        e        ,88~-_   
+  888   /   e88~~8e   e88~~/  e88~~8e  888-~88e _d88__       888   /      d8b      d888   /  
+  888    | d888  88b d888    d888  88b 888  888  888         888    |    /Y88b    88888    | 
+  888    | 8888__888 8888    8888__888 888  888  888         888    |   /  Y88b   88888    | 
+  888   /  Y888    , Y888    Y888    , 888  888  888         888   /   /____Y88b   Y888   /  
+  888_-~    "88___/   "88__/  "88___/  888  888  "88_/       888_-~   /      Y88b   '88_-~   
+  `);
+}
